Create README.md
DS2 can perform joins using hash objects similarly to the data step.  The real advantage of PROC DS2 is it can distribute the work across multiple cores unlike a data step.  The tradeoff is that each core will need:
1)	A set of unique rows from the left table
2)	A full copy of the hash object in memory
This is not a problem for inner and left joins.  We are leveraging more resources to get valuable results back quicker.  In the case of the full join, the extra condition of keeping unmatched rows from both tables does become problematic.  Since the left table is partitioned across cores, each core is not able to independently determined